+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes to this project.
 - added `check_new_round` to check if a new round has started
 - return proper Python data types, for example the NMR amounts are now
   floats and no longer strings
+- show progress bar while downloading dataset
 
 ## [0.7.1] - 2017-12-29
 - fix import issues (py2 vs py3)